+++ itest/test_collection_functions.py
@@ -17,6 +17,7 @@ from thingsdb.exceptions import OverflowError
 from thingsdb.exceptions import ZeroDivisionError
 from thingsdb.exceptions import OperationError
 from thingsdb.exceptions import ThingsDBError
+from thingsdb.exceptions import SyntaxError
 from thingsdb.client.protocol import Err
 
 
@@ -1762,6 +1763,53 @@ class TestCollectionFunctions(TestBase):
         self.assertEqual(await client.query('int(false);'), 0)
         self.assertEqual(await client.query('int("3.14");'), 3)
         self.assertEqual(await client.query('int("-3.14");'), -3)
+
+    async def test_closure(self, client):
+        with self.assertRaisesRegex(
+                LookupError,
+                'type `nil` has no function `closure`'):
+            await client.query('nil.closure();')
+
+        with self.assertRaisesRegex(
+                NumArgumentsError,
+                'function `closure` takes at most 1 argument '
+                'but 2 were given'):
+            await client.query('closure("||42", nil);')
+
+        with self.assertRaisesRegex(
+                TypeError,
+                'cannot convert type `regex` to `closure`'):
+            await client.query('closure(/.*/);')
+
+        with self.assertRaisesRegex(
+                TypeError,
+                'cannot convert type `nil` to `closure`'):
+            await client.query('closure(nil);')
+
+        with self.assertRaisesRegex(
+                LookupError,
+                'node is not a closure'):
+            await client.query('closure("nil")')
+
+        with self.assertRaisesRegex(
+                BadDataError,
+                'closure is expecting exactly one statement'):
+            await client.query('closure("||123; 42")')
+
+        with self.assertRaisesRegex(
+                SyntaxError,
+                'invalid syntax in closure'):
+            await client.query('closure("||(");')
+
+        with self.assertRaisesRegex(
+                SyntaxError,
+                'invalid syntax in closure'):
+            await client.query('closure("||");')
+
+        self.assertEqual(await client.query('closure();'), '||nil')
+        self.assertIs(await client.query('c=closure(); c();'), None)
+        self.assertEqual(await client.query('c=closure(||42); c();'), 42)
+        self.assertEqual(await client.query('c=closure(|x|x*2); c(21);'), 42)
 
     async def test_regex(self, client):
         with self.assertRaisesRegex(